api-agent/requirements.txt
python-dotenv
requests
ssl